chore: add libpolly-22-dev for bcc install

+++ vhdbuilder/scripts/linux/ubuntu/tool_installs_ubuntu.sh
@@ -30,7 +30,7 @@ installBcc() {
     if [ "${VERSION}" = "22.04" ] || [ "${VERSION}" = "24.04" ]; then
         apt_get_install 120 5 300 build-essential git bison cmake flex libedit-dev libllvm14 llvm-14-dev libclang-14-dev python3 zlib1g-dev libelf-dev libfl-dev || exit $ERR_BCC_INSTALL_TIMEOUT
     elif [ "${VERSION}" = "26.04" ]; then
-        apt_get_install 120 5 300 build-essential git bison cmake flex libedit-dev libllvm22 llvm-22-dev libclang-22-dev zlib1g-dev libelf-dev libfl-dev || exit $ERR_BCC_INSTALL_TIMEOUT
+        apt_get_install 120 5 300 build-essential git bison cmake flex libedit-dev libllvm22 llvm-22-dev libclang-22-dev python3 zlib1g-dev libelf-dev libfl-dev || exit $ERR_BCC_INSTALL_TIMEOUT
     else
         apt_get_install 120 5 300 build-essential git bison cmake flex libedit-dev libllvm6.0 llvm-6.0-dev libclang-6.0-dev python zlib1g-dev libelf-dev python3-distutils libfl-dev || exit $ERR_BCC_INSTALL_TIMEOUT
     fi
@@ -43,6 +43,11 @@ installBcc() {
     # libPolly.a is needed for the make target that runs later, which is not present in the default patch version of llvm-14 that is downloaded for 24.04
     if [ "${VERSION}" = "24.04" ]; then
       apt_get_install 120 5 300 libpolly-14-dev || exit $ERR_BCC_INSTALL_TIMEOUT
+    fi
+
+    # libPolly.a is needed for the make target that runs later, which is not present in the default patch version of llvm-22 that is downloaded for 26.04
+    if [ "${VERSION}" = "26.04" ]; then
+      apt_get_install 120 5 300 libpolly-22-dev || exit $ERR_BCC_INSTALL_TIMEOUT
     fi
 
     local bcc_version="v0.29.0"
@@ -74,6 +79,7 @@ installBcc() {
     if [ "${VERSION}" = "22.04" ] || [ "${VERSION}" = "24.04" ]; then
         apt_get_purge 120 5 300 bison cmake flex libedit-dev libllvm14 llvm-14-dev libclang-14-dev zlib1g-dev libelf-dev libfl-dev || exit $ERR_BCC_INSTALL_TIMEOUT
     elif [ "${VERSION}" = "26.04" ]; then
+        # we remove git to keep the image as minimal as possible
         apt_get_purge 120 5 300 git bison cmake flex libedit-dev libllvm22 llvm-22-dev libclang-22-dev zlib1g-dev libelf-dev libfl-dev || exit $ERR_BCC_INSTALL_TIMEOUT
     else
         apt_get_purge 120 5 300 git bison cmake flex libedit-dev libllvm6.0 llvm-6.0-dev libclang-6.0-dev zlib1g-dev libelf-dev libfl-dev || exit $ERR_BCC_INSTALL_TIMEOUT
@@ -82,6 +88,11 @@ installBcc() {
     # libPolly.a is needed for the make target that runs later, which is not present in the default patch version of llvm-14 that is downloaded for 24.04
     if [ "${VERSION}" = "24.04" ]; then
       apt_get_purge 120 5 300 libpolly-14-dev || exit $ERR_BCC_INSTALL_TIMEOUT
+    fi
+
+    # libPolly.a is needed for the make target that runs later, which is not present in the default patch version of llvm-22 that is downloaded for 26.04
+    if [ "${VERSION}" = "26.04" ]; then
+      apt_get_install 120 5 300 libpolly-22-dev || exit $ERR_BCC_INSTALL_TIMEOUT
     fi
 
     rm -rf /tmp/bcc
